#3 Make sure that 2 player works
